modified:   .DS_Store 	new file:   .package.json.swp 	renamed:    Models/conection.js -> Models/connection.js 	modified:   Models/marcacoesModels.js 	modified:   app.js 	modified:   package-lock.json 	modified:   package.json 	renamed:    public/paginaInformacaoPessoal.html -> public/adminHomePage.html 	modified:   public/index.html 	modified:   public/informacao.html 	modified:   public/javascripts/listaMarcacoes.js 	modified:   public/marcacoesRecolha.html 	modified:   public/novaMarcacao.html 	modified:   routes/marcacoesRouter.js
