+++ app/package.json
@@ -10,6 +10,8 @@
     "debug": "~2.6.9",
     "express": "~4.16.1",
     "morgan": "~1.9.1",
-    "stop": "^3.1.0"
+    "msql": "^1.0.3",
+    "stop": "^3.1.0",
+    "utils": "^0.3.1"
   }
 }
