+++ app/public/informacao.html
@@ -12,8 +12,8 @@
 			<h1 class="logotipo"><a hrefl="index.html">BloodBus</a></h1>
 			<ul>
 				<li><a href="index.html">Home</a></li>
-				<li><a href="#">Ser Dador</a></li>
-				<li><a href="oaginaInformacaoPessoal.html">Marcação</a></li>
+				<li><a href="informacao.html">Ser Dador</a></li>
+				<li><a href="adminHomePage.html">Marcação</a></li>
 				<li><a href="#">Contactos</a></li>
 			</ul>
 		</nav>
