+++ app/public/marcacoesRecolha.html
@@ -14,14 +14,15 @@
 			<ul>
 				<li><a href="index.html">Home</a></li>
 				<li><a href="informacao.html">Ser Dador</a></li>
-				<li><a href="paginaInformacaoPessoal.html">Marcação</a></li>
+				<li><a href="adminHomePage.html">Marcação</a></li>
 				<li><a href="#">Contactos</a></li>
 			</ul>
 		</nav>
 	</header>
 
 	<main>
-		
+		<ul id="listaMarcacoes">
+		</ul>	
 	</main>	
 
 	<footer>
